Fix FabTransform circular reveal calculation
This isn't really noticable at high speeds from a fab to a dialog, but slowed down you see that the transition from the fab to an activity cuts off after approx. 75% of the animation and just jumps to the non-clipped activity. With this fix the end radius gets calculated correctly. Same for the reversed transition.